ranged enemy fleeing from player if eh's too close
added a feature that makes enemy ai run away from  the player if he apporaches too close. Additionally, separated AI classes into a base class and specific enemy class and created a new ranged enemy prefab to simplyfy work.
