+++ Assets/Prefabs/enemy prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a14e0179511d0b44dabb330a64a856fb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
