+++ Assets/Prefabs/enemy prefabs/Bullet.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ae956bb906d19f248bbbcb36c6e13908
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
