+++ Assets/Prefabs/enemy prefabs/Ranged Enemy.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 22dfe26c02695f14b908b07c71d2be93
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
