+++ Assets/Prefabs/enemy prefabs/Soldier 1.prefab
@@ -25,7 +25,7 @@ Transform:
   m_GameObject: {fileID: 4537297885330694700}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0.458, y: 0.5, z: 0.781}
+  m_LocalPosition: {x: 0, y: 0.5, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -60,7 +60,7 @@ Transform:
   m_GameObject: {fileID: 5661323289070885356}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 41.8, y: 1, z: 41.4}
+  m_LocalPosition: {x: 0, y: 1, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
@@ -164,6 +164,8 @@ MonoBehaviour:
     serializedVersion: 2
     m_Bits: 64
   blood: {fileID: 0}
+  orbs: {fileID: 0}
+  healthBar: {fileID: 0}
 --- !u!1 &6152167338885103131
 GameObject:
   m_ObjectHideFlags: 0
@@ -194,7 +196,7 @@ Transform:
   m_GameObject: {fileID: 6152167338885103131}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: -41.8, y: -1, z: -41.4}
+  m_LocalPosition: {x: 0, y: -3, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
