+++ Assets/Scripts/enemyAiScripts/RangedEnemyAIScript.cs.meta
